digi-bridge/requirements.txt
aiohttp==3.6.2
pyhumps==1.3.1
APScheduler==3.6.0
-f https://s3pypi.mettel-automation.net/igzpackages/index.html
igzpackages==2.0.1
